+++ Gruntfile.js
@@ -7,6 +7,7 @@ module.exports = function(grunt) {
 
 	// Loads the different modules used by this gruntfile
 	grunt.loadNpmTasks('grunt-contrib-jshint');
+	grunt.loadNpmTasks('grunt-angular-templates');
 	grunt.loadNpmTasks('grunt-karma');
 	grunt.loadNpmTasks('grunt-protractor-runner');
 	grunt.loadNpmTasks('grunt-concurrent'); // loads concurrent runner
@@ -26,7 +27,7 @@ module.exports = function(grunt) {
 	// use this one when you're coding e2e tests
 	grunt.registerTask('e2e', ['concurrent:e2e']);
 	// this task updates all distributions - launch it once before each release
-	grunt.registerTask('dist', ["ts:dist", 'concat:spe', 'uglify:spe', 'concat:standard', 'uglify:standard', 'sass:dist']);
+	grunt.registerTask('dist', ["ts:dist", "ngtemplates:dist", 'concat:spe', 'uglify:spe', 'concat:standard', 'uglify:standard', 'sass:dist']);
 	// this updates the dists and tests it, creates karma coverage
 	grunt.registerTask('test', ['dist', "ts:test", 'karma:debug', 'karma:coverage', 'protractor:singlerun', 'jshint']);